// src/app/transfers/models/api/addedTransferOutputModel.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { TransactionCategory } from "~/app/accounts/models/api";

/**
 * Payment that is added but not yet confirmed
 */
export interface AddedTransferOutputModel {
    transactionId?: string;
    valueDate?: Date;
    amount?: number;
    amountCurrency?: string;
    currencyPair?: string;
    fxRate?: number;
    categoryCode?: TransactionCategory;
    categoryName?: string;
    destinationAccountIBAN?: string;
    sourceAccountTransactionName?: string;
    destinationAccountTransactionName?: string;
}