// src/app/transfers/models/api/confirmedTransferOutputModel.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TransactionCategory, TransactionStatus } from "~/app/accounts/models/api";

export interface ConfirmedTransferOutputModel {
    transactionId?: string;
    valueDate?: Date;
    bookedDate?: Date;
    amount?: number;
    currencyPair?: string;
    fxRate?: number;
    currencyCode?: string;
    isBooked?: boolean;
    statusName?: string;
    statusCode?: TransactionStatus;
    categoryCode?: TransactionCategory;
    categoryName?: string;
    sourceAccountTransactionName?: string;
    destinationAccountTransactionName?: string;
    transactionReference?: string;
}